Add proof-of-concept boost::asio::ssl implementation for SSPI
Add a very basic proof-of-concept for using Windows SChannel/SSPI with
boost::asio for TLS/SSL communication.

This tries to mimick how the original boost::asio OpenSSL
implementation works, but is far from complete, doesn't work in many
cases (including creating a TLS connection to boost.org) and isn't
even asynchronous.

This is only meant to demonstrate that it should be possible to create
an implementation so OpenSSL isn't required on Windows, but is far
from actually doing that.