Support building test client on non-windows platforms
Fix CMake build script to link with openssl on non-windows platforms
and add some ugly ifdefs in the very primitive https_client to alias
the ssl namespace to either the asio::ssl namespace I'm trying to
mimick or my new namespace.

Also add an unused method enum and constructor arguement to the
context constructor to mirror the interface of the original
asio::ssl::context constructor.

This is to have some easy way of comparing results/behavior during
implementation.