Generate test certificates at build time
Instead of keeping test certificates required for testing the TLS
communication in the repository, generate them at build time.

This means they will not have to be manually updated if they should
expire and avoids having to keep private keys in the repository, which
seems wrong even if it doesn't matter in this case as the private key
is not private.