Generate random data for openssl on Windowss
It seems like running openssl on Windows in appveyor requires a file
with random data to be written to the users home directory.

For some reason that doesn't seem to be the case when running locally,
even with the same version of openssl.

Just create the random file and hope that helps.