Extend unittest to make client read response
Extend the (still very basic) unittest so the client reads the data
returned from the server with data being in various sizes.

This points out a bug in the current implementation that only makes it
more clear that the logic for handling state and decryption buffers
should be moved to a seperate class, but now at least there's a test
that should hopefully make the risk of breaking it less likely.