Move encrypted message for sending to separate class
Since the structure required for encrypting a message is the only one
that actually needs the header, max size and trailer size, move it to
a seperate class for better separation of responsibilities.

This should also make it easier to refactor it to using a fixed size
constant buffer since the max size should be known.