Clean up error code
Move implementation details away from anonymous namespace (which
shouldn't be used in a header file) and into detail namespace.

Remove stream error category that was probably copy pasted from the
asio::ssl implementation but not used here anwyay.

Rename error category to sspi_category as it is our own category and
shouldn't conflict with the ssl_category which is used for openssl
errors.