Ensure correct executor is used when calling complete
If an error has been set initially by any of the m_sspi classes used
by the async helper classes it would cause completion handler (invoked
by self.complete()) to be executed on the wrong executor.

Ensure that doesn't happen by posting the completion handler instead
of calling it directly if not already continuing the a coroutine.