Use catch2 unittest framework for testing
Use catch2 for unittests as it provides a lot of useful functionality
that I prefer not having to implement myself as would be the case
using boost lightweight test.

It might make sense to avoid having an external dependency in the
longer run, but if that becomes an issue it is hopefully not too hard
to rewrite existing tests and create some of the helpful macros etc.

For no I consider it more important to focus on writing some actual
tests as easily as possible.