Avoid copy of shutdown data
Instead of copying the shutdown message to send to the server when
closing the TLS connection, use the buffer directly and add a consume
member function for releasing it once it has been sent.

This could and should most likely be generalized and used in
read/write/handshake as well.