Support Windows default certficate store(s)
Enable verifying server certificates against the default trusted CAs
in the Windows certificate store(s) when set_default_verify_paths() is
called on the context.

This should mimick the behavior of the OpenSSL implementation used in
boost::asio::ssl.

Also, support setting the verify_mode, again trying to mimick the
OpenSSL implementation although it is basically used as a boolean
here.