Move CredHandle to sppi_impl class
Since the CredHandle needs to be constructed based on whether the
handshake should be performed on the client or server side, it needs
to be moved to sspi_impl.

Also add an extra function to handshake class for setting
client/server handshake, but still only client handshake implemented
for now.