Remove verify_mode enum
Use a boolean for whether to enable server certificate
verification.

There's not any reasonable way to mimick the OpenSSL veriify behavior
anyway, so this is a move towards making the API expose the
functionality provided by this implementation while hopefully still
making this fairly simple to use together with the existing OpenSSL
stream implementation.