Add client/server wrappers for wintls and asio_ssl implementations
Instead of using way too many template arguments and unpacking a way
to big tuple with those arguments, add some wrappers that exposes the
functionality shared between the boost::asio:ssl and this wintls
implentation.

This makes the code a lot more readable. For now just used for the
general echo test.