Use CMAKE_CXX_COMPILER CMake variable to find Clang compiler
Since the appveyor image doesn't seem to have the LLVM toolset install
with the Visual Studio 2019 image, but does have LLVM installed, use
the CMake variable for specifying the C++ compiler to find Clang.

The requires calling the Visual Studio batch file for setting the
correct paths to toolsets etc.