Remove CMAKE_CXX_STANDARD
MSVC only support setting the C++ standard to C++14 or newer and using
CMAKE_CXX_STANDARD for setting it to anything else actually enforces
using the latest standard which is quite surprising.

At least for now, just remove it. It probably makes more sense to
require specific compiler support instead which currently is the
latest version of Clang and MSVC.