Use CERT_CONTEXT type for adding certificate authorities
Use the CERT_CONTEXT type native for Windows for adding certificates
used for validating server certificates when doing client TLS
handshakes.

Move the conversion of standard X509 certificates to CERT_CONTEXT type
to a separate function so it is still easy to use in parallel with the
asio::ssl implementation.