Comment out crashing test section
The Windows API often likes to see if something interesting might be
at the end of null pointers even if the documentation states
otherwise.

Uncomment the crashing test for invalid data until I figure out a way
to generate an invalid certificate that isn't so invalid that it make
the test executable crash.