Split documentation sections into seperate pages
Instead of having one huge file with the documentation including
examples and what is generated from the sources, split it into
separate sections to improve navigation a bit.

This also means using the fulltoc extension to expand the section
currently being viewed.

This is not exactly how I want it, but still better than it was.