Cleanup CMake files
It doesn't really seem to make much sense to attempt to build any of
the code on other platforms than Windows. At least for now. So cleanup
the CMake files so only the documentation is generated on sane
platforms and the code is built on Windows. This also means removing a
few ugly #ifdefs.