Limit copy of decrypted data
Copy as much as possible of the decrypted data to the use supplied
buffer directly when decrypting/reading data.

A buffer is still needed in case the user supplied buffer cannot hold
all the decrypted data.

It might be possible to optimize this further by using the size of the
user supplied buffer as a hint for the size of the buffer to use for
decryption making it much more likely that a temporary buffer is not
needed at all. On the other hand that would potentially require more
calls to decrypt which might be an even bigger overhead, and there's
most likely a minimum size required for the buffer passed to the
DecryptMessage function.

Any further optimizations would require a bit of profiling, this is an
obvious improvement.