Use fixed sized buffer for decrypted data
Since we know that there will never be more decrypted data than data
for encryption, a fixed sized buffer can be used to store the
decrypted data that doesn't fit in the user supplied buffer.

Apart from avoiding to resize the buffer to fit the data it more
importantly means not having to move the contents around since it's
fairly trivial to keep track of the remaining data inside the fixed
buffer.