Don't enable warnings as errors globally
Although warnings as error is generally a very good idea, enabling
them for a library is not very friendly as it means everyone using the
library will also get warnings as errors which might not be what they
want.

Disable that for the library, but enable it for the unittests so the
build will still fail on any warning during test.