Use CERT_CONTEXT for server certificate in public interface
Instead of trying to copy the interface from asio::ssl directly, which
cannot really be done since the implementations differ too much, just
use a native windows CERT_CONTEXT as the certificate to use for a
server stream.

This makes the interface simpler and gives the user the flexibility to
add and use certificates and keys in the way that is native for the
platform.

Move the code for interacting with standard formats for keys and
certificates to utility functions instead.