Use bootstrap theme for documentation
Use the sphinx bootstrap theme for documentation and add a fairly
hackish sphinx extension for adding the toctree as a dictionary to the
HTML render context that can then be used to generate a custom navbar.

This looks much better than how it was previously and good enough that
I can now focus on writing code, documenting it and with the extension
also restructure the documents without worrying about the layout.

Hopefully...