Make stream movable
Instead of trying to make each member of the stream class movable move
the stream members to a separate class and store that in a unique_ptr.

That makes the code quite a bit simpler, actually reduces the amount
of heap allocations and most importantly actually works as some
sub-members of the stream are not trivially moveable.