Use CPack for generating release package
Instead of manually creating the .zip package for releases use CPack
which will just copy whatever is in the source folder.

In order to release the generated HTML documentation as well, merge
the documentation and release jobs and move the generated HTML
documentation to the source folder.

That means that any changes to the documentation will only be uploaded
when a new release is made but that makes much more sense anyway.