Use Windows headers
Instead of trying to redefine Windows types just include windows.h.

It would definitely be nice to avoid Windows headers as that hurts
compile times and defines macros that can cause hard to debug errors,
but it's not very simple to do so for now just live with it.

This also means that this can be build with UNICODE defined, although
this is deliberately not being used for the SSPI headers as that would
require some ugly #ifdefs and the only place we actually pass a string
is for a hostname which doesn't allow UNIOCODE characters anyway.