Cleanup CI jobs
Create different job configs for different configurations and run
those instead of trying to create a matrix with all possible
configurations.

It doesn't make much sense to create both a debug and release build
for the mingw Makefile generator for example as the build type is
defined at configure time.

This should also make it simpler to add more jobs with different
build/config options.