Pass Hostname to Certificate Verification
Certificates are not validated by means of the sspi context.
Instead they are validated manually using CertVerifyCertificateChainPolicy.
Therefore, we pass the host name to this function so it can verify it.

Prior to this change, connections to hosts presenting a valid certificate
with an incorrect Common Name were made.
Example: https://wrong.host.badssl.com/