Add OCSP Test
Revocation checking was introduced in #63,
but the tests only cover CRLs so far.
This makes use of the simple OCSP responder included
in OpenSSL to add a testcase covering OCSP.

The dependency on boost::filesystem comes through boost::process.

Also generate_certs.sh was simplified a bit to store all generated files
in a subfolder 'gen'.