Refactor unittest CI jobs
Instead of defining all options for each job in the matrix use a
combination of Boost versions, MS toolsets, build type and compiler.

That should hopefully make it simpler to extend or change and greatly
increases the platform coverage.

Currently explicitly limited a bit to avoid having way to many CI
jobs.