Upgrade install-boost github action to 2.3.1
Upgrade the install-boost action to 2.3.1. This means we can set the
toolset being used (msvc or mingw) to avoid a warning.

The actual toolset used for compiling boost doesn't matter for now as
this project only uses boost headers but it might not be the case in
the future and is the correct thing to do anyway.