Increase the read size in small reads test
For some reason some update to some part of Windows broke TLS
handshakes some times when using smaller reads with an error of
SEC_E_ILLEGAL_MESSAGE returned after some calls to
InitializeSecurityContext during a client handshake.

It is not clear currently what broke this or whether this really
matters. Since there doesn't seem to be any workaround and assuming
this is just another bug in Windows TLS code that might not mean
anything outside of this test just increase the size being used in the
test.