Do not delete Move Operations of Context
context Members are bool, enum, detail::context_certificates
detail::context_certificates has a bool and two unique_ptr.
Move operations are not problematic for any of these.

This does, for example, allow user code to have a factory functions
to set up a context in a certain way and return it.