Clarify use of private certificates when operating as server
In Windows certificates can have an associated private key which seems
to be known as "private certificates". Clarify that private
certificates needs to be used when setting the certificate to be used
as a server and test that the keys match (or the private key is
missing) when adding setting the certificate to be used when operating
as a server.