add shutdown test cases
This should not be merged in the current state but is rather supposed
to show of a difference between Asio.SSL and WinTLS:
Asio will produce a stream_truncated error on shutdown, if the remote
did not properly shut down the protocol while WinTLS will not.

Note https://stackoverflow.com/questions/25587403/boost-asio-ssl-async-shutdown-always-finishes-with-an-error

Also, according to that post, Asio should produce an error if it initiated the
shutdown and it goes through without problem.
This does not seem to be the case (anymore?).
This is only relevant for the https examples in WinTLS.
Currently the https async example handles asio::error::eof
(as do the beast examples), while the sync example does not.