Move manual_auth out of the handshake loop.
Or rather move it to the end where it is called at most once.
This resolves the TODO on sspi_handshake::state
and is intended to simplify the code to allow for further refactoring.

There is a small change in behavior:
As a server, all communication with the client is now done before manual_auth.
Before, manual_auth was called as early as possible and
possibly more data was sent through the stream after that.