Don't install OpenSSL in GitHub actions Windows runner
The Windoes images used by GitHub actions now ships with OpenSSL
development files (headers and libs) so no reason to install OpenSSL
as a package.

Not only is it not required but it actually causes inconsistencies
between the headers used when compiling and the libraries used at
runtime causing various extremely confusing runtime errors.