Fetch thirdparty test dependencies with FetchContent
Instead of copying thirdparty implementations of variant and
string_view directly into this repository fetch them with CMake
FetchContent only when targeting C++14 without boost which is when
they are actually required.

This avoids cluttering up this repository and makes it easier to
distinguish between our code and thirdparty code as well as updating
that.