Remove external subprocess code
Using a thirdparty library just for starting and stopping a process is
a bit overkill so just use the Win32 API functions for that.

Declare the funtions and structs needed in the file needed them to
avoid including Windows headers.

If some more functionality related to process handling is needed a
dedicated process library (eg. Boost.Process) can be introduced.