Fix test stream naming
Move the test stream code to test_stream directory instead of a very
generic utils directory since it now only contains code related to the
test stream.

The test stream is still more or less a copy of the same stream from
Boost.Beast test code with a few additional headers as needed.

Update the header guards to reflect that this code belongs here and
change the end of namespace comments to be consistent with the rest of
the code in this repository.