Move library out of boost namespace
This library is now called asio.wintls instead of boost.wintls as it
no longer depends on boost.

Since this library does not require boost remove the boost namespace
and simply use the wintls namespace.

Update the documentation to reflect that change and refer to asio
instead of boost asio where it makes most sense.