Link examples to currently released version
Instead of linking to the examples in the master branch, fetch the
version tag passed by command line when building the documentation.

Getting the version from the command line by looking at the command
line args in the conf.py file is a bit of a hack but there doesn't
seem to be a cleaner way to do this and the documentation is always
intended to be build with CMake anyway.