Discover test when building with mingw
For some reason the list of tests will be empty when using a wildcard
for the test spec when built with mingw. We want to use a wildcard in
order to include the integration tests. For some reason prepending a
space makes it work in all cases.

No reason to dig further into this before upgrading to catch v3.