Silence GCC warning on null dereference
Boost.Beast includes directly on indirectly uses some boost algorithm
function that makes GCC warns on some potential null
dereference. Since this function gets inlined it will generate an
warning even if the header is a system include so we need to silence
that.

For some reason this cannot be silenced with the usual -Wno- option
but telling the optimizer not to delete null pointer checks
works. This can hopefully be removed in a later version of GCC.